script/requirements.txt
addict==2.4.0
attrs==23.2.0
blinker==1.7.0
cattrs==23.2.3
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.1.0
coremltools==7.1
cryptography==43.0.0
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dtaidistance==2.3.12
exceptiongroup==1.2.0
fastjsonschema==2.19.1
ffmpeg==1.3
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.42.0
fsspec==2024.2.0
future==0.18.3
gcloud==0.18.3
googleapis-common-protos==1.64.0
h264decoder @ file:///Users/daehwakim/Desktop/research/ProjTrack/core/h264decoder
httplib2==0.22.0
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.1
importlib_resources==6.4.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jwcrypto==1.5.6
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.42.0
lz4==4.3.2
MarkupSafe==2.1.5
matplotlib==3.7.2
mpmath==1.3.0
nbformat==5.5.0
nest-asyncio==1.6.0
networkx==3.2.1
npy-append-array==0.9.16
numba==0.59.1
numpy==1.25.2
oauth2client==4.1.3
open3d==0.16.0
opencv-contrib-python==4.9.0.80
opencv-python==3.4.17.63
packaging==23.1
pandas==2.2.1
Pillow==10.0.0
platformdirs==4.2.0
plotly==5.19.0
protobuf==3.20.3
py4j==0.10.9.7
pyaml==23.12.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.5
PyBoof==0.43.1
pycparser==2.22
pycryptodome==3.20.0
pygame==2.5.2
pyliblzfse==0.4.1
pyparsing==3.0.9
pyquaternion==0.9.9
Pyrebase4==4.8.0
python-dateutil==2.8.2
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
pyzlib==0.1.16
referencing==0.33.0
requests==2.31.0
requests-toolbelt==0.10.1
retrying==1.3.4
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scikit-video==1.1.11
scipy==1.11.1
six==1.16.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
traitlets==5.14.1
transforms3d==0.4.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.19
websockets==13.0.1
Werkzeug==3.0.1
zipp==3.17.0
